+++ Frends.AzureDataLake.UploadFiles/Frends.AzureDataLake.UploadFiles/Definitions/Options.cs
@@ -1,4 +1,5 @@
 using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 
 namespace Frends.AzureDataLake.UploadFiles.Definitions;
 
@@ -8,10 +9,19 @@ namespace Frends.AzureDataLake.UploadFiles.Definitions;
 public class Options
 {
     /// <summary>
-    /// True: Throw an exception.
-    /// False: If the error is ignorable, such as when a File already exists, the error will be added to the Result.ErrorMessages list instead of stopping the Task.
+    /// True: Throw an exception on failure.
+    /// False: Return Result with Success=false and Error with description of exception.
     /// </summary>
     /// <example>true</example>
     [DefaultValue(true)]
     public bool ThrowErrorOnFailure { get; set; } = true;
+
+    /// <summary>
+    /// Custom error message to use when ThrowErrorOnFailure is true or when returning an error result.
+    /// If empty, the original exception message is used.
+    /// </summary>
+    /// <example></example>
+    [DisplayFormat(DataFormatString = "Text")]
+    [DefaultValue("")]
+    public string ErrorMessageOnFailure { get; set; } = string.Empty;
 }